requirements.txt
dill==0.3.3
pandas==1.2.4
plotly==4.14.3
praw==7.2.0
python-dotenv==0.17.0
requests==2.25.1
